requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
Pweave==0.23
PyHamcrest==1.8.1
Pygments==2.0.1
Pymacs==0.25
Sphinx==1.2.3
alabaster==0.6.2
argparse==1.2.1
behave==1.2.4
-e hg+ssh://hg@bitbucket.org/rallion/cameraobscura@28de78795c5e55a5eff57f025485ac5b3a3257eb#egg=cameraobscura-testsuites
configobj==5.0.6
docopt==0.6.2
docutils==0.12
ecdsa==0.11
enum34==1.0.4
iperflexer==2014.12.22.1
matplotlib==1.4.2
mock==1.0.1
nose==1.3.4
numpy==1.9.1
paramiko==1.15.2
parse==1.6.6
parse-type==0.3.4
pudb==2014.1
pycrypto==2.6.1
pyparsing==2.0.3
python-dateutil==2.3
pytz==2014.10
six==1.8.0
sphinxcontrib-plantuml==0.5
-e hg+ssh://hg@bitbucket.org/cloisteredmonkey-admin/theape@ae625720ecec016ec57b913be2b856c24e9f503b#egg=theape-dev
urwid==1.3.0
wsgiref==0.1.2
